+++ SWP291_Team2_SE1829NJ/src/java/controller/room/roomCreate.java
@@ -3,7 +3,8 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/JSP_Servlet/Servlet.java to edit this template
  */
 
-package controller.manager;
+package controller.room;
+
 
 import java.io.IOException;
 import java.io.PrintWriter;
