+++ SWP291_Team2_SE1829NJ/src/java/controller/room/roomEdit.java
@@ -3,7 +3,8 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/JSP_Servlet/Servlet.java to edit this template
  */
 
-package controller.manager;
+package controller.room;
+
 
 import java.io.IOException;
 import java.io.PrintWriter;
@@ -16,7 +17,7 @@ import jakarta.servlet.http.HttpServletResponse;
  *
  * @author HUNG
  */
-public class roomManage extends HttpServlet {
+public class roomEdit extends HttpServlet {
    
     /** 
      * Processes requests for both HTTP <code>GET</code> and <code>POST</code> methods.
@@ -33,10 +34,10 @@ public class roomManage extends HttpServlet {
             out.println("<!DOCTYPE html>");
             out.println("<html>");
             out.println("<head>");
-            out.println("<title>Servlet roomManage</title>");  
+            out.println("<title>Servlet roomEdit</title>");  
             out.println("</head>");
             out.println("<body>");
-            out.println("<h1>Servlet roomManage at " + request.getContextPath () + "</h1>");
+            out.println("<h1>Servlet roomEdit at " + request.getContextPath () + "</h1>");
             out.println("</body>");
             out.println("</html>");
         }
